# backend/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plan2Field AI",
    "version": "0.1.0",
    "description": "Local prototype. All actual mutations require project review. No external scheduling writes."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/projects/{projectId}/proposals": {
      "get": {
        "summary": "List project proposals",
        "operationId": "get__projects_projectId_proposals",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ]
      }
    },
    "/api/projects/{projectId}/proposals/{proposalId}": {
      "get": {
        "summary": "Get scoped proposal",
        "operationId": "get__projects_projectId_proposals_proposalId",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "patch": {
        "summary": "Revise and revalidate candidate selection",
        "operationId": "patch__projects_projectId_proposals_proposalId",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token from /auth/me. Browser Origin must equal APP_ORIGIN."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "proposal_version",
                  "selected_activity_id",
                  "proposed_changes"
                ],
                "properties": {
                  "proposal_version": {
                    "type": "integer"
                  },
                  "selected_activity_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "proposed_changes": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/proposals/{proposalId}/validate": {
      "post": {
        "summary": "Preview server-validated actuals without committing",
        "operationId": "post__projects_projectId_proposals_proposalId_validate",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token from /auth/me. Browser Origin must equal APP_ORIGIN."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "selected_activity_id",
                  "proposed_changes"
                ],
                "properties": {
                  "selected_activity_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "proposed_changes": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/proposals/{proposalId}/approve": {
      "post": {
        "summary": "Reviewer approve",
        "operationId": "post__projects_projectId_proposals_proposalId_approve",
        "responses": {
          "200": {
            "description": "Successful response; reports/imports use asynchronous processing"
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Project role or CSRF check failed"
          },
          "409": {
            "description": "Stale version or idempotency conflict"
          },
          "422": {
            "description": "Business-rule validation failed"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token from /auth/me. Browser Origin must equal APP_ORIGIN."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalVersion"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "Event": {
        "additionalProperties": false,
        "properties": {
          "event_type": {
            "enum": [
              "START",
              "FINISH",
              "PROGRESS",
              "BLOCKER",
              "PLAN",
              "CORRECTION",
              "UNKNOWN"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "event_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Event Date"
          },
          "date_basis": {
            "default": "EXPLICIT",
            "title": "Date Basis",
            "type": "string"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Action"
          },
          "line_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Line Number"
          },
          "asset_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Asset Tag"
          },
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Area"
          },
          "discipline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Discipline"
          },
          "quantity": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Unit"
          },
          "quantity_mode": {
            "anyOf": [
              {
                "enum": [
                  "INCREMENTAL",
                  "CUMULATIVE",
                  "COMPONENT_SET"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Quantity Mode"
          },
          "component_ids": {
            "items": {
              "type": "string"
            },
            "title": "Component Ids",
            "type": "array"
          },
          "negated": {
            "default": false,
            "title": "Negated",
            "type": "boolean"
          },
          "future_intent": {
            "default": false,
            "title": "Future Intent",
            "type": "boolean"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "title": "Missing Fields",
            "type": "array"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "title": "Blockers",
            "type": "array"
          },
          "correction_of_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Correction Of Event Id"
          }
        },
        "required": [
          "event_type"
        ],
        "title": "Event",
        "type": "object"
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "ProposalVersion": {
        "type": "object",
        "required": [
          "proposal_version"
        ],
        "properties": {
          "proposal_version": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
    }
  }
}